requirements.txt
aiohttp
asyncio
discord.py
requests
discord
python-dotenv
PyYAML
youtube-transcript-api
PyPDF2
python-docx
tqdm
pyshorteners
yt_dlp
fake-useragent
pillow
openai
PyNaCl
psutil
spotipy
pymongo
geopy
imagehash
gifmaker
opencv-python
scikit-image
fastapi
youtube_dl
bs4